Minor fixes during code review

+++ Coursera/Yandex_CppWhite/Coursera_CppWhite_Week01_Task03.cpp
@@ -1,20 +1,17 @@
 #include <iostream>
 
 int divide()
-
 {
    using namespace std;
    int a, b;
    cin >> a >> b;
    if (b == 0)
-
    {
       cout << "Impossible";
    }
    else
-
    {
-      int r = a / float(b);
+      float r = a / float(b);
       cout << r;
    }
    cout << endl;
